CustomUserDetails id의 타입를 Long에서 Integer로 변경 수정하였습니다. UserAddressDTO.java의 primaryAddressId, secondlyAddressId의 타입을 Integer->String 으로 수정하였습니다.

+++ src/main/java/org/devkirby/hanimman/config/CustomUserDetails.java
@@ -17,7 +17,7 @@ public class CustomUserDetails implements UserDetails {
     private Instant createdAt;
     private Instant deletedAt;
     private Collection<? extends GrantedAuthority> authorities;
-    private long id;
+    private Integer id;
     private Address primaryAddressId;
 
 
@@ -43,7 +43,7 @@ public class CustomUserDetails implements UserDetails {
         return null;
     }
 
-    public Long getId(){
+    public Integer getId(){
         return id;
     }
